+++ src/createReducerContext.ts
@@ -1,11 +1,11 @@
-import { createFactory, createContext, useContext, useReducer } from 'react';
+import { createElement, createContext, useContext, useReducer } from 'react';
 
 const createReducerContext = <R extends React.Reducer<any, any>>(
   reducer: R,
   defaultInitialState: React.ReducerState<R>
 ) => {
   const context = createContext<[React.ReducerState<R>, React.Dispatch<React.ReducerAction<R>>] | undefined>(undefined);
-  const providerFactory = createFactory(context.Provider);
+  const providerFactory = (props, children) => createElement(context.Provider, props, children);
 
   const ReducerProvider: React.FC<{ initialState?: React.ReducerState<R> }> = ({ children, initialState }) => {
     const state = useReducer<R>(reducer, initialState !== undefined ? initialState : defaultInitialState);