+++ src/createStateContext.ts
@@ -1,8 +1,8 @@
-import { createFactory, createContext, useContext, useState } from 'react';
+import { createElement, createContext, useContext, useState } from 'react';
 
 const createStateContext = <T>(defaultInitialValue: T) => {
   const context = createContext<[T, React.Dispatch<React.SetStateAction<T>>] | undefined>(undefined);
-  const providerFactory = createFactory(context.Provider);
+  const providerFactory = (props, children) => createElement(context.Provider, props, children);
 
   const StateProvider: React.FC<{ initialValue?: T }> = ({ children, initialValue }) => {
     const state = useState<T>(initialValue !== undefined ? initialValue : defaultInitialValue);