+++ src/useDrop.ts
@@ -1,8 +1,4 @@
-/* eslint-disable */
-import * as React from 'react';
-import useMountedState from './useMountedState';
-
-const { useState, useMemo, useCallback, useEffect } = React;
+import { useState, useMemo, useCallback, useEffect } from 'react';
 
 export interface DropAreaState {
   over: boolean;
@@ -23,13 +19,8 @@ export interface DropAreaOptions {
 }
 
 const noop = () => {};
-/*
-const defaultState: DropAreaState = {
-  over: false,
-};
-*/
 
-const createProcess = (options: DropAreaOptions, mounted: boolean) => (dataTransfer: DataTransfer, event) => {
+const createProcess = (options: DropAreaOptions) => (dataTransfer: DataTransfer, event) => {
   const uri = dataTransfer.getData('text/uri-list');
 
   if (uri) {
@@ -42,29 +33,26 @@ const createProcess = (options: DropAreaOptions, mounted: boolean) => (dataTrans
     return;
   }
 
-  if (dataTransfer.items && dataTransfer.items.length) {
-    dataTransfer.items[0].getAsString(text => {
-      if (mounted) {
-        (options.onText || noop)(text, event);
-      }
-    });
+  if (event.clipboardData) {
+    const text = event.clipboardData.getData('text');
+    (options.onText || noop)(text, event);
+    return;
   }
 };
 
 const useDrop = (options: DropAreaOptions = {}, args = []): DropAreaState => {
   const { onFiles, onText, onUri } = options;
-  const isMounted = useMountedState();
   const [over, setOverRaw] = useState<boolean>(false);
   const setOver = useCallback(setOverRaw, []);
-  const process = useMemo(() => createProcess(options, isMounted()), [onFiles, onText, onUri]);
+  const process = useMemo(() => createProcess(options), [onFiles, onText, onUri]);
 
   useEffect(() => {
-    const onDragOver = event => {
+    const onDragOver = (event) => {
       event.preventDefault();
       setOver(true);
     };
 
-    const onDragEnter = event => {
+    const onDragEnter = (event) => {
       event.preventDefault();
       setOver(true);
     };
@@ -77,13 +65,13 @@ const useDrop = (options: DropAreaOptions = {}, args = []): DropAreaState => {
       setOver(false);
     };
 
-    const onDrop = event => {
+    const onDrop = (event) => {
       event.preventDefault();
       setOver(false);
       process(event.dataTransfer, event);
     };
 
-    const onPaste = event => {
+    const onPaste = (event) => {
       process(event.clipboardData, event);
     };