+++ src/useGetSetState.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import { useCallback, useRef } from 'react';
 import useUpdate from './useUpdate';