+++ src/useIntersection.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import { RefObject, useEffect, useState } from 'react';
 
 const useIntersection = (