+++ src/useLockBodyScroll.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import { RefObject, useEffect, useRef } from 'react';
 
 export function getClosestBody(el: Element | HTMLElement | HTMLIFrameElement | null): HTMLElement | null {
@@ -46,49 +45,68 @@ let documentListenerAdded = false;
 export default !doc
   ? function useLockBodyMock(_locked: boolean = true, _elementRef?: RefObject<HTMLElement>) {}
   : function useLockBody(locked: boolean = true, elementRef?: RefObject<HTMLElement>) {
-      elementRef = elementRef || useRef(doc!.body);
+      const bodyRef = useRef(doc!.body);
+      elementRef = elementRef || bodyRef;
+
+      const lock = (body) => {
+        const bodyInfo = bodies.get(body);
+        if (!bodyInfo) {
+          bodies.set(body, { counter: 1, initialOverflow: body.style.overflow });
+          if (isIosDevice) {
+            if (!documentListenerAdded) {
+              document.addEventListener('touchmove', preventDefault, { passive: false });
+
+              documentListenerAdded = true;
+            }
+          } else {
+            body.style.overflow = 'hidden';
+          }
+        } else {
+          bodies.set(body, { counter: bodyInfo.counter + 1, initialOverflow: bodyInfo.initialOverflow });
+        }
+      };
+
+      const unlock = (body) => {
+        const bodyInfo = bodies.get(body);
+        if (bodyInfo) {
+          if (bodyInfo.counter === 1) {
+            bodies.delete(body);
+            if (isIosDevice) {
+              body.ontouchmove = null;
+
+              if (documentListenerAdded) {
+                document.removeEventListener('touchmove', preventDefault);
+                documentListenerAdded = false;
+              }
+            } else {
+              body.style.overflow = bodyInfo.initialOverflow;
+            }
+          } else {
+            bodies.set(body, { counter: bodyInfo.counter - 1, initialOverflow: bodyInfo.initialOverflow });
+          }
+        }
+      };
 
       useEffect(() => {
         const body = getClosestBody(elementRef!.current);
         if (!body) {
           return;
         }
-
-        const bodyInfo = bodies.get(body);
-
         if (locked) {
-          if (!bodyInfo) {
-            bodies.set(body, { counter: 1, initialOverflow: body.style.overflow });
-            if (isIosDevice) {
-              if (!documentListenerAdded) {
-                document.addEventListener('touchmove', preventDefault, { passive: false });
-
-                documentListenerAdded = true;
-              }
-            } else {
-              body.style.overflow = 'hidden';
-            }
-          } else {
-            bodies.set(body, { counter: bodyInfo.counter + 1, initialOverflow: bodyInfo.initialOverflow });
-          }
+          lock(body);
         } else {
-          if (bodyInfo) {
-            if (bodyInfo.counter === 1) {
-              bodies.delete(body);
-              if (isIosDevice) {
-                body.ontouchmove = null;
-
-                if (documentListenerAdded) {
-                  document.removeEventListener('touchmove', preventDefault);
-                  documentListenerAdded = false;
-                }
-              } else {
-                body.style.overflow = bodyInfo.initialOverflow;
-              }
-            } else {
-              bodies.set(body, { counter: bodyInfo.counter - 1, initialOverflow: bodyInfo.initialOverflow });
-            }
-          }
+          unlock(body);
         }
       }, [locked, elementRef.current]);
+
+      // clean up, on un-mount
+      useEffect(() => {
+        const body = getClosestBody(elementRef!.current);
+        if (!body) {
+          return;
+        }
+        return () => {
+          unlock(body);
+        };
+      }, []);
     };