+++ src/useMediatedState.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import { Dispatch, SetStateAction, useCallback, useRef, useState } from 'react';
 
 export interface StateMediator<S = any> {