+++ src/usePrevious.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef } from 'react';
 
-const usePrevious = <T>(state: T): T | undefined => {
+export default function usePrevious<T>(state: T): T | undefined {
   const ref = useRef<T>();
 
   useEffect(() => {
@@ -8,6 +8,4 @@ const usePrevious = <T>(state: T): T | undefined => {
   });
 
   return ref.current;
-};
-
-export default usePrevious;
+}