+++ src/useSearchParam.ts
@@ -1,11 +1,11 @@
-/* eslint-disable */
 import { useState, useEffect } from 'react';
 
 const getValue = (search: string, param: string) => new URLSearchParams(search).get(param);
 
 export type UseQueryParam = (param: string) => string | null;
 
-const useSearchParam: UseQueryParam = param => {
+const useSearchParam: UseQueryParam = (param) => {
+  const location = window.location;
   const [value, setValue] = useState<string | null>(() => getValue(location.search, param));
 
   useEffect(() => {