+++ src/useUnmountPromise.ts
@@ -12,10 +12,10 @@ const useUnmountPromise = (): Race => {
     const race = <P extends Promise<any>, E>(promise: P, onError?: (error: E) => void) => {
       const newPromise: P = new Promise((resolve, reject) => {
         promise.then(
-          result => {
+          (result) => {
             if (!refUnmounted.current) resolve(result);
           },
-          error => {
+          (error) => {
             if (!refUnmounted.current) reject(error);
             else if (onError) onError(error);
             else console.error('useUnmountPromise', error);